% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/body_get_game_sim_game_sim_post.R
\docType{class}
\name{BodyGetGameSimGameSimPost}
\alias{BodyGetGameSimGameSimPost}
\title{BodyGetGameSimGameSimPost}
\format{
An \code{R6Class} generator object
}
\description{
BodyGetGameSimGameSimPost Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{game}}{\link{Game}}

\item{\code{start_state}}{Start State \link{State} [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{BodyGetGameSimGameSimPost$new()}}
\item \href{#method-toJSON}{\code{BodyGetGameSimGameSimPost$toJSON()}}
\item \href{#method-fromJSON}{\code{BodyGetGameSimGameSimPost$fromJSON()}}
\item \href{#method-toJSONString}{\code{BodyGetGameSimGameSimPost$toJSONString()}}
\item \href{#method-fromJSONString}{\code{BodyGetGameSimGameSimPost$fromJSONString()}}
\item \href{#method-clone}{\code{BodyGetGameSimGameSimPost$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BodyGetGameSimGameSimPost$new(game, start_state = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BodyGetGameSimGameSimPost$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BodyGetGameSimGameSimPost$fromJSON(BodyGetGameSimGameSimPostJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BodyGetGameSimGameSimPost$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BodyGetGameSimGameSimPost$fromJSONString(BodyGetGameSimGameSimPostJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BodyGetGameSimGameSimPost$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
