% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\docType{class}
\name{Player}
\alias{Player}
\title{Player}
\format{
An \code{R6Class} generator object
}
\description{
Player Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{full_name}}{Full Name character}

\item{\code{position}}{Position character}

\item{\code{mlb_id}}{Mlb Id integer}

\item{\code{debut_date}}{Debut Date character}

\item{\code{team_id}}{Team Id integer}

\item{\code{team_abbrev}}{Team Abbrev character}

\item{\code{active}}{Active character}

\item{\code{bat_side}}{Bat Side character}

\item{\code{pitch_hand}}{Pitch Hand character}

\item{\code{birth_country}}{Birth Country character}

\item{\code{birth_date}}{Birth Date character}

\item{\code{current_age}}{Current Age integer}

\item{\code{photo_url}}{Photo Url character}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Player$new()}}
\item \href{#method-toJSON}{\code{Player$toJSON()}}
\item \href{#method-fromJSON}{\code{Player$fromJSON()}}
\item \href{#method-toJSONString}{\code{Player$toJSONString()}}
\item \href{#method-fromJSONString}{\code{Player$fromJSONString()}}
\item \href{#method-clone}{\code{Player$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$new(
  full_name,
  position,
  mlb_id,
  debut_date,
  team_id,
  team_abbrev,
  active,
  bat_side,
  pitch_hand,
  birth_country,
  birth_date,
  current_age,
  photo_url,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$fromJSON(PlayerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$fromJSONString(PlayerJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Player$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
