% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state.R
\docType{class}
\name{State}
\alias{State}
\title{State}
\format{
An \code{R6Class} generator object
}
\description{
State Class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{on_1b}}{On 1B character [optional]}

\item{\code{on_2b}}{On 2B character [optional]}

\item{\code{on_3b}}{On 3B character [optional]}

\item{\code{inning}}{Inning integer [optional]}

\item{\code{outs}}{Outs integer [optional]}

\item{\code{top}}{Top character [optional]}

\item{\code{bat_score}}{Bat Score integer [optional]}

\item{\code{fld_score}}{Fld Score integer [optional]}

\item{\code{bat_lineup_order}}{Bat Lineup Order integer [optional]}

\item{\code{fld_lineup_order}}{Fld Lineup Order integer [optional]}

\item{\code{pitch_number}}{Pitch Number integer [optional]}

\item{\code{frame_runs_scored}}{Frame Runs Scored integer [optional]}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{State$new()}}
\item \href{#method-toJSON}{\code{State$toJSON()}}
\item \href{#method-fromJSON}{\code{State$fromJSON()}}
\item \href{#method-toJSONString}{\code{State$toJSONString()}}
\item \href{#method-fromJSONString}{\code{State$fromJSONString()}}
\item \href{#method-clone}{\code{State$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$new(
  on_1b = FALSE,
  on_2b = FALSE,
  on_3b = FALSE,
  inning = 1,
  outs = 0,
  top = TRUE,
  bat_score = 0,
  fld_score = 0,
  bat_lineup_order = 1,
  fld_lineup_order = 1,
  pitch_number = 0,
  frame_runs_scored = 0,
  ...
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSON"></a>}}
\if{latex}{\out{\hypertarget{method-toJSON}{}}}
\subsection{Method \code{toJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$toJSON()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSON"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSON}{}}}
\subsection{Method \code{fromJSON()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$fromJSON(StateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-toJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-toJSONString}{}}}
\subsection{Method \code{toJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$toJSONString()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-fromJSONString"></a>}}
\if{latex}{\out{\hypertarget{method-fromJSONString}{}}}
\subsection{Method \code{fromJSONString()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$fromJSONString(StateJson)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{State$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
